requirements.txt
Django>=5.1.*
psycopg2-binary>=2.9
djangorestframework>=3.15
python-dotenv==1.0.1